refactor: split to crate
